requirements.txt
boto3==1.16.31
celery==5.0.3
celery-redbeat==2.0.0
decorator==4.4.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.4
# django-anymail[sendinblue]==7.1.0
django-cache-memoize==0.1.7
django-cors-headers
django-debug-toolbar
django-extensions==3.1.0
django-filter==2.4.0
django-money==1.2.1
django-redis-cache==3.0.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-request==1.5.6
django-storages==1.10.1
django-timezone-field
django-user-agents==0.4.0
django-widget-tweaks
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
graphviz
ipython
ipywidgets
markdownify
numpy==1.18.0 # newer version doesn't play well with py3.8, macos
pandas==1.1.3 # newer version doesn't play well with py3.8, macos
plotly
python-telegram-bot==13.1
pytz==2020.4
PyYAML
redis==3.5.3
requests-oauthlib==1.3.0
requests==2.25.0
schedule==0.6.0
simplejson
six
toolz
waitress==1.4.4

# FINANCE
ccxt==1.38.94
TA-Lib==0.4.19
backtrader
yahoo_fin
alpha_vantage # for TA indicators
quandl # for frequent price calls
polygon-api-client  # for frequent price calls


# LOCAL DEV
psycopg2-binary

# PRODUCTION ONLY
# psycopg2
